+++ src/test/java/com/treasuredata/android/TreasureDataTest.java
@@ -164,7 +164,7 @@ public class TreasureDataTest extends TestCase {
     public void testAddEventAndUploadCustomEventBlocked() throws IOException {
         Map<String, Object> records = new HashMap<String, Object>();
         records.put("key", "val");
-        when(td.isCustomEventBlocked()).thenReturn(true);
+        when(td.isCustomEventDisabled()).thenReturn(true);
         td.addEvent("db_", "tbl", records);
         td.uploadEvents();
         assertEquals(0, client.addedEvent.size());
@@ -174,7 +174,7 @@ public class TreasureDataTest extends TestCase {
         Map<String, Object> records = new HashMap<String, Object>();
         records.put("key", "val");
         records.put("__is_auto_track_event", "true");
-        when(td.isCustomEventBlocked()).thenReturn(true);
+        when(td.isCustomEventDisabled()).thenReturn(true);
         td.addEvent("db_", "tbl", records);
         td.uploadEvents();
         assertEquals(1, client.addedEvent.size());
@@ -183,7 +183,7 @@ public class TreasureDataTest extends TestCase {
     public void testAddEventAndUploadCustomEventUnBlocked() throws IOException {
         Map<String, Object> records = new HashMap<String, Object>();
         records.put("key", "val");
-        when(td.isCustomEventBlocked()).thenReturn(false);
+        when(td.isCustomEventDisabled()).thenReturn(false);
         td.addEvent("db_", "tbl", records);
         td.uploadEvents();
         assertEquals(1, client.addedEvent.size());