+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Change Log
 
 ## Version 0.1.18
-_2018-04-26_
+_2018-05-03_
 
-* Auto event tracking is off by default
-* Add TreasureData#enableTrackAppLifecycleEvents(String table)
+* Auto event tracking is optional and off by default
+* Add TreasureData#trackAppLifecycleEvents(String table)
 * Add TreasureData#resetUUID(String table)
 * Add functions for auto tracking events and custom event opt out
 * Add Opt Out example for td-android-sdk-demo